requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
pillow>=10.0.0
matplotlib>=3.7.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
mistralai>=1.0.0
